BGDIINF_SB-1404: * moved update_temporal_extent method and subroutines in separate file * added extra parameter and datetimes to debug logging * renamed methods for clearness * removed redundant if statements
